+++ lib/dradis-pdf_export.rb
@@ -3,6 +3,7 @@ require 'dradis-plugins'
 
 # Load supporting classes
 require 'prawn'
+require 'prawn/table'
 
 # Load this add-on's engine
 require "dradis/plugins/pdf_export"